Modified import sequence for App and Greeting components
